requirements.txt
# SecureTalk - Dependencies

# Flask and extensions
Flask==2.2.5
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6
Flask-Limiter==3.5.0
Werkzeug==2.2.3

# Security
cryptography==41.0.5
WTForms==3.1.1
email-validator==2.1.0

# WebSockets
python-socketio==5.10.0
python-engineio==4.8.0

# Environment variables
python-dotenv==1.0.0

# Production server
gunicorn==21.2.0